requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# API and web framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0

# UI
streamlit>=1.10.0
plotly>=5.3.0
matplotlib>=3.4.0

# F1 data
fastf1>=2.2.0

# AI/ML integrations
openai>=0.27.0
groq>=0.3.2

# Utilities
python-dotenv>=0.19.0
joblib>=1.1.0